Add translations & restructure server
