+++ api/setup.cfg
@@ -1,2 +1,11 @@
 [flake8]
 max_line_length = 150
+
+[isort]
+include_trailing_comma = true
+multi_line_output = 3
+lines_after_imports = 2
+combine_as_imports = true
+force_grid_wrap = 0
+use_parentheses = true
+line_length = 88
