+++ api/src/db.py
@@ -1,0 +1,16 @@
+import sqlite3 as sql
+
+from .app import app
+
+
+DB_PATH = app.config.get("DB_PATH", "database.sqlite3")
+
+
+@app.before_first_request
+def create_db_schema():
+    with sql.connect(DB_PATH) as con:
+        cur = con.cursor()
+        cur.execute(
+            "CREATE TABLE IF NOT EXISTS mail_addresses(mail_address text PRIMARY KEY NOT NULL, timestamp DATETIME DEFAULT CURRENT_TIMESTAMP);"
+        )
+        con.commit()
