+++ api/src/templates/confirmation-email.jinja
