+++ api/src/templates/metadata.jinja
