+++ scripts/extract-translations.sh
@@ -5,7 +5,7 @@ cd "${0%/*}/.."
 
 # extract translations
 npx ngx-translate-extract -i ./src -o ./i18n/templates/client.pot --clean --format pot -m _
-pybabel extract -F api/babel.cfg -o ./i18n/templates/server.pot --no-location ./api
+pybabel extract -F api/babel.cfg -o ./i18n/templates/server.pot ./api
 
 # merge translations
 echo "merging POT files..."
