+++ src/app/prices/prices.component.html
@@ -40,9 +40,7 @@
         <div class="price-col blue">
             <div class="price-col-header">
                 <h2 translate>Tagung</h2>
-                <p translate>
-                    für mittelgroße (virtuelle) Versammlungen
-                </p>
+                <p translate>für mittelgroße (virtuelle) Versammlungen</p>
             </div>
             <div class="price-col-body">
                 <div class="price-col-price">
@@ -69,9 +67,7 @@
         <div class="price-col grey">
             <div class="price-col-header">
                 <h2 translate>Kongress</h2>
-                <p translate>
-                    für große (virtuelle) Konferenzen
-                </p>
+                <p translate>für große (virtuelle) Konferenzen</p>
             </div>
             <div class="price-col-body">
                 <div class="price-col-price">
@@ -152,11 +148,11 @@
             <li id="audio">
                 <u><b translate>Audiokonferenz – ideal zum Austausch für kleine virtuelle Gremiensitzung</b></u>
                 <br>
-                <span translate>
-                    Verwenden Sie OpenSlides mit integrierter Audiokonferenz. Über unseren exklusiv für OpenSlides betriebenen Jitsi-Server kommunizieren alle Teilnehmende direkt miteinander - geschützt in einem Passwort-gesicherten Audio-Konferenzraum. Jede/r verbindet sich automatisch nach dem Login in OpenSlides mit der Audiokonferenz und darf das eigene Mikrofon ein- und ausschalten.
+                <span>
+                    {{ "Verwenden Sie OpenSlides mit integrierter Audiokonferenz. Über unseren exklusiv für OpenSlides betriebenen Jitsi-Server kommunizieren alle Teilnehmende direkt miteinander - geschützt in einem Passwort-gesicherten Audio-Konferenzraum. Jede/r verbindet sich automatisch nach dem Login in OpenSlides mit der Audiokonferenz und darf das eigene Mikrofon ein- und ausschalten." | translate }}
                 </span>
-                <a href="https://files.openslides.org/press/3.2/OpenSlides3.2-Audiokonferenz.gif" translate>
-                    » Mehr erfahren ...
+                <a href="https://files.openslides.org/press/3.2/OpenSlides3.2-Audiokonferenz.gif">
+                    » {{ "Mehr erfahren" | translate }} ...
                 </a>
                 <br>
                 <b><i>50 € / {{ 'Monat' | translate }} · {{ 'bis 50 Teilnehmende' | translate }}</i></b>
@@ -164,15 +160,15 @@
         
             <li id="video">
                 <u><b translate>Video-Livestream – für große virtuelle Versammlungen und Parteitage</b></u><br>
-                <span translate>
-                    Integrieren Sie einen Video-Livestream Ihrer virtuellen Veranstaltung in OpenSlides. Alle Teilnehmenden verfolgen den Stream jederzeit sichtbar in OpenSlides. Möchte sich eine Person zu Wort melden, setzt sich diese selbständig auf die aktuelle Redeliste in OpenSlides und wechselt damit automatisch in einen Jitsi-Videokonferenzraum. Der Livestream überträgt die Videokonferenz. Nach Ende der Wortmeldung verlässt man den Konferenzraum und verfolgt wieder den Livestream.
+                <span>
+                    {{ "Integrieren Sie einen Video-Livestream Ihrer virtuellen Veranstaltung in OpenSlides. Alle Teilnehmenden verfolgen den Stream jederzeit sichtbar in OpenSlides. Möchte sich eine Person zu Wort melden, setzt sich diese selbständig auf die aktuelle Redeliste in OpenSlides und wechselt damit automatisch in einen Jitsi-Videokonferenzraum. Der Livestream überträgt die Videokonferenz. Nach Ende der Wortmeldung verlässt man den Konferenzraum und verfolgt wieder den Livestream." | translate }}
                 </span>
                 <br>
-                <span translate>
-                    Dieses Verfahren ist sehr bandbreitenschonend (im Vergleich zu reinen Videokonferenzen) und erlaubt eine große Skalierung an Teilnehmerzahlen. Virtuelle Veranstaltungen mit mehreren hundert Personen sind damit kein Problem mehr. Hinzu kommt, dass die meisten Teilnehmenden eine Veranstaltung nur "konsumierend" verfolgen, d.h. nur ein kleiner Personenkreis möchte sich aktiv mit Mikrofon und Kamera zu Wort melden. Setzen Sie auf erprobte Broadcast-Streaming-Lösung – intelligent gekoppelt mit der OpenSlides-Redelistenfunktion.
+                <span>
+                    {{ 'Dieses Verfahren ist sehr bandbreitenschonend (im Vergleich zu reinen Videokonferenzen) und erlaubt eine große Skalierung an Teilnehmerzahlen. Virtuelle Veranstaltungen mit mehreren hundert Personen sind damit kein Problem mehr. Hinzu kommt, dass die meisten Teilnehmenden eine Veranstaltung nur "konsumierend" verfolgen, d.h. nur ein kleiner Personenkreis möchte sich aktiv mit Mikrofon und Kamera zu Wort melden. Setzen Sie auf erprobte Broadcast-Streaming-Lösung – intelligent gekoppelt mit der OpenSlides-Redelistenfunktion.' | translate }}
                 </span>
-                <a href="https://files.openslides.org/press/3.2/OpenSlides3.2-Videokonferenz.gif" translate>
-                    » Mehr erfahren ...
+                <a href="https://files.openslides.org/press/3.2/OpenSlides3.2-Videokonferenz.gif">
+                    » {{ "Mehr erfahren" | translate }} ...
                 </a>
                 <br>
                 <span translate>Kontaktieren Sie uns! Wir finden die passende Streaming-Lösung für Ihre Veranstaltung.</span>
@@ -181,8 +177,8 @@
                 
             <li id="saml">
                 <u><b translate>Single Sign-On via SAML – für Organisationen mit zentralem Mitgliedsserver</b></u><br>
-                <span translate>
-                    Ermöglichen Sie Ihren Mitgliedern sich mit ihren bekannten Mitgliedsaccounts in OpenSlides anzumelden. OpenSlides unterstützt den Standard "SAML" (Security Assertion Markup Language) für Single Sign-On (SSO). Ersparen Sie sich damit die Teilnehmerverwaltung in OpenSlides. Und Ihre Mitglieder brauchen sich keinen extra OpenSlides-Login merken.
+                <span>
+                    {{ 'Ermöglichen Sie Ihren Mitgliedern sich mit ihren bekannten Mitgliedsaccounts in OpenSlides anzumelden. OpenSlides unterstützt den Standard "SAML" (Security Assertion Markup Language) für Single Sign-On (SSO). Ersparen Sie sich damit die Teilnehmerverwaltung in OpenSlides. Und Ihre Mitglieder brauchen sich keinen extra OpenSlides-Login merken.' | translate }}
                 </span>
                 <br><b><i>50 € / {{ 'Monat' | translate }}</i></b>
             </li>
@@ -197,8 +193,8 @@
             Sind noch Fragen offen? Haben Sie individuelle Wünsche? Schreiben Sie uns an <a class="inline" href="mailto:info@openslides.com">info@openslides.com</a>!
         </p>
         <p>
-            <span translate>
-                OpenSlides ist Freie Software (Open Source) und kann – mit einigen administrativen Kenntnissen – auch selbständig auf dem eigenen Server installiert werden.
+            <span>
+                {{ "OpenSlides ist Freie Software (Open Source) und kann – mit einigen administrativen Kenntnissen – auch selbständig auf dem eigenen Server installiert werden." | translate }}
             </span>
             <span>&nbsp;<a class="inline" routerLink="/" fragment="getting-started">{{ 'Mehr erfahren' | translate }}...</a></span>
         </p>
